MARL-task/requirements.txt
# Core dependencies
numpy>=1.21.0
torch>=2.0.0
gymnasium>=0.28.0

# Data processing
pandas>=1.5.0
pyyaml>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Utilities
tqdm>=4.64.0
tensorboard>=2.12.0
python-json-logger>=2.0.7

# Development and testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Optional: for better performance
scipy>=1.9.0

# Web Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
